Tested Grid Colliders
- Put colliders on the tiles of the grid. Successfully detected
collision between test player character and tile grids. Will utilized
OnTriggerEnter2D and OnTriggerExit2D to tell the tiles when a unit is
occupying them or not.

- Updated the grid so it no longer spawns objects at runtime. Instead,
All the grids have been statically created in the editor so that we can
better design levels.

- Changed grid cell dimensions to 120x120 as per Jose's recommendation
since they fit the 16:9 aspect ratio better.

- Updated the building and unit names in the Tile class.

- GridManager still creates a 2D array of Tiles and assigns them x, y
values at runtime. It iterates over the children of the grid game object
and assigns them accordingly.

- Added trigger colliders to the GridButton prefab. These will be used
to detect units and buildings that are on them.

- Tile now inherits from MonoBehaviour. This is necessary so we can use
OnTriggerEnter2D and OnTriggerExit2D. We could have the units dictate
this logic, but for the sake of decoupling code, I think it would be
best for the tile to handle recognition of units that are on it.

- GridButton prefab now has Tile script attached to it. x and y values
are exposed in the inspector, if we need to see them for any reason.