Updated Grid Editor Tools
Grid now has a script attached to it that will allow you to resize the cell sizing and maintain tile positioning relative to their old positions.

This tool will help us decide the size we want for our tiles.

When the grid cell sizes are updated, the tile snapping is updated to match the dimensions.